fixed ErrorText bug on TimetableScreen
